// Random-cats/src/types/ApiResponse.ts
// Weight object
export interface Weight {
  imperial: string;
  metric: string;
}

// Cat details object
export interface CatDetails {
  id: number;
  name: string;
  vetstreet_url: string;
  temperament: string;
  origin: string;
  country_codes: string;
  country_code: string;
  description: string;
  life_span: string;
  indoor: number;
  lap: number;
  alt_names: string;
  adaptability: number;
  affection_level: number;
  child_friendly: number;
  dog_friendly: number;
  energy_level: number;
  grooming: number;
  health_issues: number;
  intelligence: number;
  shedding_level: number;
  social_needs: number;
  stranger_friendly: number;
  vocalisation: number;
  experimental: number;
  hairless: number;
  natural: number;
  rare: number;
  rex: number;
  suppressed_tail: number;
  short_legs: number;
  wikipedia_url: string;
  hypoallergenic: number;
  image: string;
  weight: Weight;
}

// Outer API response
export interface ApiResponse {
  statusCode: number;
  success: boolean;
  message: string;
  data: CatDetails;
}